feat(customer): update contracts, ViewModels, and screens
Contracts:
- Add missing states/events (address history, food categories, search recent/suggest)
- Remove searchDebounced query from SearchContract
- Add deliveryFee, merchantAddress to CartContract
- Add onRatingClick, onReorderClick to OrderHistoryContract
- Remove deliveryType from OrderHistoryItem
- Expand OrderStatusFilter to 6 entries with label/matches()
- Add address-related events to EditProfileContract

ViewModels:
- Add address autocomplete, food category, favorites, search recent/suggest flows
- Add delete cart item, update address use cases
- Fix SecurityViewModel Load handler (was empty)
- Fix OrderHistoryViewModel status mapping (1:1 with backend) + multi-item title
- Fix NotifViewModel deprecated confirmValueChange migration

Screens:
- Remove deprecated Icons.Default.TrendingUp (use AutoMirrored.Filled)
- Fix CafeScreen search, category, address flows
- Fix CartScreen with updated contract
- Fix ChatScreen with ChatMessage model
- Fix EditProfileScreen with address management
- Fix FavoriteScreen with new params
- Fix NotifScreen LaunchedEffect for dismissState
- Fix OrderHistoryScreen filter/timeline/colors, remove deliveryType InfoRow
- Fix ProfileScreen with updated flows
- Fix SearchScreen with trending/suggest/recent search

+++ feature-customer/src/main/java/com/mtv/app/shopme/feature/customer/ui/ProfileScreen.kt
@@ -106,10 +106,10 @@ fun ProfileScreen(
         menuSummary?.completed ?: 0
     ).any { it > 0 }
     val orderMenus = listOf(
-        OrderMenu("Belum Bayar", Icons.Filled.AccountBalanceWallet, menuSummary?.unpaid ?: 0, "ORDERED"),
-        OrderMenu("Dipesan", Icons.Filled.Inventory, menuSummary?.ordered ?: activeOrderCount.takeUnless { hasMenuSummary } ?: 0, "ORDERED"),
-        OrderMenu("Dimasak", Icons.Filled.Inventory, menuSummary?.cooking ?: 0, "COOKING"),
-        OrderMenu("Dikirim", Icons.Filled.LocalShipping, menuSummary?.shipping ?: 0, "DELIVERING"),
+        OrderMenu("Belum Bayar", Icons.Filled.AccountBalanceWallet, (menuSummary?.unpaid ?: 0L).toInt(), "ORDERED"),
+        OrderMenu("Dipesan", Icons.Filled.Inventory, (menuSummary?.ordered ?: (activeOrderCount.takeUnless { hasMenuSummary } ?: 0).toLong()).toInt(), "ORDERED"),
+        OrderMenu("Dimasak", Icons.Filled.Inventory, (menuSummary?.cooking ?: 0L).toInt(), "COOKING"),
+        OrderMenu("Dikirim", Icons.Filled.LocalShipping, (menuSummary?.shipping ?: 0L).toInt(), "DELIVERING"),
     )
 
     when (val customerState = state.customer) {
